server/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fernando Bryan Reza Campos", email = "fer.rezac@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "prism-py>=0.0.1rc1",
    "uvicorn>=0.34.0"
]
